requirements.txt
discord.py==0.16.8
git+git://github.com/Rapptz/discord.py.git
pillow
imgurpython
boto3